Inform user if search returned 0 results
This commit fixes the issue where the "Contacting host: .." messages stays
hanging in the echo area whether or not a query ended up with non-zero result
pods.

Now after all the pods are retrieved and displayed in the WolframAlpha buffer,
the echo area is cleared out.. so the "Contacting host: .." message doesn't stay
hanging.

If the query results in 0 retrieved pods, a "No results for your query." is
printed in the WolframAlpha buffer, so that the user doesn't have to keep on
waiting for something to show up.

* wolfram--delete-in-progress-notification: New function. The logic for searching backward for the "(searching)" notification and removing
  it is moved to this function.

* wolfram--append-pods-to-buffer: Use the above function, remove the unused
  `buffer` argument.

* wolfram--query-callback: New function. Moved out the lambda callback function
as this separate function for better readability and "debuggablity".