Inform user if search returned 0 results
This commit fixes the issue where the "Contacting host: .." messages stays
hanging in the echo area whether or not a query ended up with non-zero result
pods.

Now after all the pods are retrieved and displayed in the WolframAlpha buffer,
the echo area is cleared out.. so the "Contacting host: .." message doesn't stay
hanging.

If the query results in 0 retrieved pods, this a message "Your WolframAlpha
query turned up with 0 results." is shown to the user, so that the user doesn't
have to keep on waiting for something to show up.

Also moved out the lambda callback function into a defined
`wolfram--query-callback` function for better readability and "debuggablity".